+++ DataLayerProjects/PostgresDataLayer/PostGresData.cs
@@ -1,63 +1,46 @@
-﻿using LiteDB;
-using SpoonCMS.Classes;
-using SpoonCMS.Exceptions;
-using SpoonCMS.Interfaces;
+﻿using SpoonCMSCore.Classes;
+using SpoonCMSCore.Exceptions;
+using SpoonCMSCore.Interfaces;
 using System;
 using System.Collections.Generic;
-using System.IO;
 
-namespace SpoonCMS.DataClasses
+using Marten;
+using Marten.Services;
+using System.Linq;
+
+namespace SpoonCMSCore.PostGresData
 {
-    public class LiteDBData : ISpoonData
+    public class PostGresData : ISpoonData
     {
-
+        private DocumentStore _store { get; }
         private string _connString { get; set; }
-        private string _defaultDir { get; set; } = @"Data\DB\";
-        private string _dbFileName = "SpoonData.db";
 
-        public LiteDBData()
-        {            
-            if (!Directory.Exists(_defaultDir))
-            {
-                Directory.CreateDirectory(_defaultDir);
-            }
-            _connString = _defaultDir + _dbFileName;
-        }
+        public PostGresData(string connString)
+        {        
+            _connString = connString;
 
-        /// <summary>
-        /// A path to the directory to store the DB
-        /// </summary>
-        /// <param name="connString"></param>
-        public LiteDBData(string connString)
-        {
-            if (!string.IsNullOrEmpty(connString))
+            _store = DocumentStore.For(x =>
             {
-                if (!Directory.Exists(connString))
-                {
-                    Directory.CreateDirectory(connString);
-                }
-                _connString = connString + _dbFileName;
-            }
-            else
-            {
-                if (!Directory.Exists(_defaultDir))
-                {
-                    Directory.CreateDirectory(_defaultDir);
-                }
-                _connString = _defaultDir + _dbFileName;
-            }            
-        }
+                x.Connection(_connString);
+                x.Serializer(new JsonNetSerializer { EnumStorage = EnumStorage.AsInteger });
+
+                x.Schema.For<Container>().Duplicate(y => y.Id, configure: y => y.IsUnique = true);
+
+                x.AutoCreateSchemaObjects = AutoCreate.CreateOrUpdate;
+            });
+        }        
 
         public Container GetContainer(string conName)
         {
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    Container existingCon = containers.FindOne(x => x.Name.Equals(conName));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Name == conName)
+                                            .FirstOrDefault();
 
-                    return existingCon;
+                    return container;
                 }
             }
             catch (Exception)
@@ -70,12 +53,13 @@ namespace SpoonCMS.DataClasses
         {
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    Container existingCon = containers.FindOne(x => x.Id.Equals(conId));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Id == conId)
+                                            .FirstOrDefault();
 
-                    return existingCon;
+                    return container;
                 }
             }
             catch (Exception)
@@ -84,16 +68,16 @@ namespace SpoonCMS.DataClasses
             }
         }
 
-        public List<ContainerSkinny> GetAllContainers()
+        public List<ContainerSkinny> GetAllContainers() //Need to make this lighterweight, this could get heavy
         {
             List<ContainerSkinny> retList = new List<ContainerSkinny>();
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    
-                    foreach(Container con in containers.FindAll())
+                    var containers = session.Query<Container>();
+
+                    foreach (Container con in containers)
                     {
                         retList.Add(con.GetSkinny());
                     }
@@ -111,19 +95,21 @@ namespace SpoonCMS.DataClasses
         {
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    var existingCon = containers.FindOne(x => x.Name.Equals(container.Name));
+                    bool isFound = session.Query<Container>()
+                                        .Where(x => x.Name == container.Name)
+                                        .Any();
 
-                    if (existingCon == null)
+                    if (!isFound)
                     {
-                        containers.Insert(container);
+                        session.Store<Container>(container);
+                        session.SaveChanges();
                     }
                     else
+                    {
                         throw new NameExistsException("A container with that name already exists");
-
-                    containers.EnsureIndex(x => x.Name);
+                    }
                 }
             }
             catch(Exception)
@@ -136,10 +122,12 @@ namespace SpoonCMS.DataClasses
         {
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    var existingCon = containers.FindOne(x => x.Id.Equals(container.Id));
+                    var existingCon = session.Query<Container>()
+                                            .Where(x => x.Id == container.Id)
+                                            .FirstOrDefault();
+
 
                     if (existingCon == null)
                     {
@@ -151,8 +139,8 @@ namespace SpoonCMS.DataClasses
                         existingCon.Active = container.Active;
                         existingCon.Items = container.Items;
 
-                        containers.Update(container);
-                        containers.EnsureIndex(x => x.Name);
+                        session.Update<Container>(existingCon);
+                        session.SaveChanges();
                     }                    
                 }
             }
@@ -166,10 +154,21 @@ namespace SpoonCMS.DataClasses
         {
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using(var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    containers.Delete((Query.EQ("Name", conName)));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Name == conName)
+                                            .FirstOrDefault();
+
+                    if (container != null)
+                    {
+                        session.Delete(container);
+                        session.SaveChanges();
+                    }
+                    else
+                    {
+                        throw new ItemDoesNotExistException("Container not found");
+                    }
                 }
             }
             catch(Exception)
@@ -183,17 +182,19 @@ namespace SpoonCMS.DataClasses
             try
             {
 
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    Container container = containers.FindOne(x => x.Name.Equals(conName));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Name == conName)
+                                            .FirstOrDefault();
 
                     if (container != null)
                     {
                         if(!container.Items.ContainsKey(item.Name))
                         {
                             container.Items.Add(item.Name, item);
-                            containers.Update(container);
+                            session.Update<Container>(container);
+                            session.SaveChanges();
                         }    
                         else
                         {
@@ -217,10 +218,11 @@ namespace SpoonCMS.DataClasses
             try
             {
 
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    Container container = containers.FindOne(x => x.Name.Equals(conName));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Name == conName)
+                                            .FirstOrDefault();
 
                     if (container != null)
                     {
@@ -231,7 +233,8 @@ namespace SpoonCMS.DataClasses
                             container.Items[item.Name].EndDate = item.EndDate;
                             container.Items[item.Name].Value = item.Value;
 
-                            containers.Update(container);
+                            session.Update<Container>(container);
+                            session.SaveChanges();
                         }
                         else
                         {
@@ -255,16 +258,18 @@ namespace SpoonCMS.DataClasses
             try
             {
 
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    Container container = containers.FindOne(x => x.Name.Equals(conName));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Name == conName)
+                                            .FirstOrDefault();
 
                     if (container != null)
                     {
                         container.Items.Remove(itemName);
 
-                        containers.Update(container);
+                        session.Update<Container>(container);
+                        session.SaveChanges();
                     }
                     else
                     {
@@ -282,20 +287,25 @@ namespace SpoonCMS.DataClasses
         {
             try
             {
-                using (var db = new LiteDatabase(_connString))
+                using (var session = _store.LightweightSession())
                 {
-                    var containers = db.GetCollection<Container>("Containers");
-                    Container container = containers.FindOne(x => x.Id.Equals(conId));
+                    Container container = session.Query<Container>()
+                                            .Where(x => x.Id == conId)
+                                            .FirstOrDefault();
 
                     if (container != null)
                     {
                         if (container.Name != conName) // Don't do anything, it's the same name we already have
                         {
-                            var existingCon = containers.FindOne(x => x.Name.Equals(conName)); //Make sure name isn't already taken
+                            var existingCon = session.Query<Container>()   //Make sure name isn't already taken
+                                                .Where(x => x.Name == conName)
+                                                .FirstOrDefault(); 
+
                             if (existingCon == null)
                             {
                                 container.Name = conName;
-                                containers.Update(container);
+                                session.Update<Container>(container);
+                                session.SaveChanges();
                             }
                             else
                             {
