+++ SpoonCMS.sln
@@ -3,13 +3,23 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpoonCMS", "SpoonCMS\SpoonCMS.csproj", "{AF5B6286-00D2-4E73-B54F-DCEB1449D64A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleCore", "ExampleCore\ExampleCore.csproj", "{31C67066-DCCF-45A9-9FC3-1EB2A4F5E703}"
+	ProjectSection(ProjectDependencies) = postProject
+		{7562A24A-FE8C-4703-9EC8-1513677DC1EA} = {7562A24A-FE8C-4703-9EC8-1513677DC1EA}
+		{096FBAB2-9028-4C73-BA19-09C1922B75BF} = {096FBAB2-9028-4C73-BA19-09C1922B75BF}
+	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{9389C868-71D4-410C-B645-CE115C30B756}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpoonCMSTests", "SpoonCMSTests\SpoonCMSTests.csproj", "{096FBAB2-9028-4C73-BA19-09C1922B75BF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpoonCMSTests", "SpoonCMSTests\SpoonCMSTests.csproj", "{096FBAB2-9028-4C73-BA19-09C1922B75BF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LiteDBDataLayer", "DataLayerProjects\LiteDBDataLayer\LiteDBDataLayer.csproj", "{95719D32-BE9E-442E-B042-45A52AC1CEEB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpoonCMSCore", "SpoonCMSCore\SpoonCMSCore.csproj", "{7562A24A-FE8C-4703-9EC8-1513677DC1EA}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DataLayers", "DataLayers", "{FBFB1D8B-EB8B-4163-9AC1-52311F697AB5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PostgresDataLayer", "DataLayerProjects\PostgresDataLayer\PostgresDataLayer.csproj", "{35A8AF1A-CEF5-4BC1-AC41-F37A370D5B4A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +27,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AF5B6286-00D2-4E73-B54F-DCEB1449D64A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AF5B6286-00D2-4E73-B54F-DCEB1449D64A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AF5B6286-00D2-4E73-B54F-DCEB1449D64A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AF5B6286-00D2-4E73-B54F-DCEB1449D64A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{31C67066-DCCF-45A9-9FC3-1EB2A4F5E703}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{31C67066-DCCF-45A9-9FC3-1EB2A4F5E703}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{31C67066-DCCF-45A9-9FC3-1EB2A4F5E703}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -29,12 +35,26 @@ Global
 		{096FBAB2-9028-4C73-BA19-09C1922B75BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{096FBAB2-9028-4C73-BA19-09C1922B75BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{096FBAB2-9028-4C73-BA19-09C1922B75BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95719D32-BE9E-442E-B042-45A52AC1CEEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95719D32-BE9E-442E-B042-45A52AC1CEEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95719D32-BE9E-442E-B042-45A52AC1CEEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95719D32-BE9E-442E-B042-45A52AC1CEEB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7562A24A-FE8C-4703-9EC8-1513677DC1EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7562A24A-FE8C-4703-9EC8-1513677DC1EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7562A24A-FE8C-4703-9EC8-1513677DC1EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7562A24A-FE8C-4703-9EC8-1513677DC1EA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35A8AF1A-CEF5-4BC1-AC41-F37A370D5B4A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35A8AF1A-CEF5-4BC1-AC41-F37A370D5B4A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35A8AF1A-CEF5-4BC1-AC41-F37A370D5B4A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35A8AF1A-CEF5-4BC1-AC41-F37A370D5B4A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{096FBAB2-9028-4C73-BA19-09C1922B75BF} = {9389C868-71D4-410C-B645-CE115C30B756}
+		{95719D32-BE9E-442E-B042-45A52AC1CEEB} = {FBFB1D8B-EB8B-4163-9AC1-52311F697AB5}
+		{35A8AF1A-CEF5-4BC1-AC41-F37A370D5B4A} = {FBFB1D8B-EB8B-4163-9AC1-52311F697AB5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F5001995-F010-4B4A-91FD-3E1B6DD903F7}
