--- SpoonCMS/Classes/NotFoundContentItem.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SpoonCMS.Classes
-{
-    public class NotFoundContentItem : ContentItem
-    {
-    }
-}
