+++ SpoonCMSCore/Classes/ServiceResponse.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpoonCMS.Classes
+namespace SpoonCMSCore.Classes
 {
     public class ServiceResponse<T>
     {
