+++ SpoonCMSCore/Files/Templates/ContainerDetails.hbs
