+++ SpoonCMSCore/Interfaces/ISpoonData.cs
@@ -1,9 +1,9 @@
-﻿using SpoonCMS.Classes;
+﻿using SpoonCMSCore.Classes;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpoonCMS.Interfaces
+namespace SpoonCMSCore.Interfaces
 {
     public interface ISpoonData
     {
