+++ SpoonCMSCore/bundleconfig.json
