+++ SpoonCMSTests/SpoonCMSTests.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
 
     <IsPackable>false</IsPackable>
   </PropertyGroup>
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SpoonCMS\SpoonCMS.csproj" />
+    <ProjectReference Include="..\SpoonCMSCore\SpoonCMSCore.csproj" />
   </ItemGroup>
 
 </Project>
